cleaned up test-calls
