+++ monitor/monitor_test.go
@@ -10,31 +10,11 @@ import (
 	"github.com/dedis/cothority/log"
 )
 
-// setupMonitor launches a basic monitor with a created Stats object
-// When finished with the monitor, just call `End()`
-func setupMonitor(t *testing.T) (*Monitor, *Stats) {
-	log.TestOutput(testing.Verbose(), 2)
-	m := make(map[string]string)
-	m["servers"] = "1"
-	stat := NewStats(m)
-	// First set up monitor listening
-	mon := NewMonitor(stat)
-	go mon.Listen()
-	time.Sleep(100 * time.Millisecond)
-
-	// Then measure
-	err := ConnectSink("localhost:" + strconv.Itoa(mon.SinkPort))
-	EnableMeasure(true)
-	if err != nil {
-		t.Fatal(fmt.Sprintf("Error starting monitor: %s", err))
-	}
-	return mon, stat
+func TestMain(m *testing.M) {
+	log.MainTest(m)
 }
 
 func TestReadyNormal(t *testing.T) {
-	defer log.AfterTest(t)
-
-	log.TestOutput(testing.Verbose(), 3)
 	m := make(map[string]string)
 	m["servers"] = "1"
 	stat := NewStats(m)
@@ -65,9 +45,6 @@ func TestReadyNormal(t *testing.T) {
 }
 
 func TestKeyOrder(t *testing.T) {
-	defer log.AfterTest(t)
-
-	log.TestOutput(testing.Verbose(), 3)
 	m := make(map[string]string)
 	m["servers"] = "1"
 	m["hosts"] = "1"
@@ -92,4 +69,24 @@ func TestKeyOrder(t *testing.T) {
 	if !bytes.Equal(str.Bytes(), str2.Bytes()) {
 		t.Fatal("KeyOrder / output not the same for same stats")
 	}
+}
+
+// setupMonitor launches a basic monitor with a created Stats object
+// When finished with the monitor, just call `End()`
+func setupMonitor(t *testing.T) (*Monitor, *Stats) {
+	m := make(map[string]string)
+	m["servers"] = "1"
+	stat := NewStats(m)
+	// First set up monitor listening
+	mon := NewMonitor(stat)
+	go mon.Listen()
+	time.Sleep(100 * time.Millisecond)
+
+	// Then measure
+	err := ConnectSink("localhost:" + strconv.Itoa(mon.SinkPort))
+	EnableMeasure(true)
+	if err != nil {
+		t.Fatal(fmt.Sprintf("Error starting monitor: %s", err))
+	}
+	return mon, stat
 }
