+++ protocols/medco/deterministic_switching_test.go
@@ -13,9 +13,7 @@ import (
 )
 
 func TestDeterministicSwitching(t *testing.T) {
-	defer log.AfterTest(t)
 	local := sda.NewLocalTest()
-	log.TestOutput(testing.Verbose(), 1)
 	_, entityList, tree := local.GenTree(5, false, true, true)
 	sda.ProtocolRegisterName("DeterministicSwitchingTest", NewDeterministicSwitchingTest)
 	defer local.CloseAll()
