+++ protocols/medco/key_switching_test.go
@@ -14,9 +14,7 @@ import (
 )
 
 func TestKeySwitching(t *testing.T) {
-	defer log.AfterTest(t)
 	local := sda.NewLocalTest()
-	log.TestOutput(testing.Verbose(), 1)
 	_, entityList, tree := local.GenTree(5, false, true, true)
 
 	defer local.CloseAll()
