+++ protocols/medco/private_aggregate_test.go
@@ -24,8 +24,6 @@ var groupingAttrC = GroupingAttributes{grpattr1, grpattr2}
 
 //TestPrivateAggregate tests private aggregate protocol
 func TestPrivateAggregate(t *testing.T) {
-	defer log.AfterTest(t)
-	log.TestOutput(testing.Verbose(), 1)
 	local := sda.NewLocalTest()
 	_ /*entityList*/, _, tree := local.GenTree(10, false, true, true)
 	sda.ProtocolRegisterName("PrivateAggregateTest", NewPrivateAggregateTest)
