+++ protocols/medco/probabilistic_switching_test.go
@@ -5,7 +5,6 @@ import (
 	"testing"
 	"time"
 
-	"github.com/dedis/cothority/log"
 	"github.com/dedis/cothority/network"
 	"github.com/dedis/cothority/protocols/medco"
 	"github.com/dedis/cothority/sda"
@@ -15,9 +14,7 @@ import (
 
 // TestProbabilisticSwitching tests probabilistic switching protocol
 func TestProbabilisticSwitching(t *testing.T) {
-	defer log.AfterTest(t)
 	local := sda.NewLocalTest()
-	log.TestOutput(testing.Verbose(), 1)
 	_, entityList, tree := local.GenTree(5, false, true, true)
 	sda.ProtocolRegisterName("ProbabilisticSwitchingTest", NewProbabilisticSwitchingTest)
