+++ sda/overlay_test.go
@@ -3,7 +3,6 @@ package sda_test
 import (
 	"testing"
 
-	"github.com/dedis/cothority/log"
 	"github.com/dedis/cothority/network"
 	"github.com/dedis/cothority/sda"
 )
@@ -28,9 +27,6 @@ func (po *ProtocolOverlay) Release() {
 }
 
 func TestOverlayDone(t *testing.T) {
-	defer log.AfterTest(t)
-
-	log.TestOutput(testing.Verbose(), 4)
 	// setup
 	h1 := sda.NewLocalHost(2000)
 	defer h1.Close()
