+++ sda/processor_test.go
@@ -20,7 +20,6 @@ func init() {
 }
 
 func TestProcessor_AddMessage(t *testing.T) {
-	defer log.AfterTest(t)
 	p := NewServiceProcessor(nil)
 	log.ErrFatal(p.RegisterMessage(procMsg))
 	if len(p.functions) != 1 {
@@ -48,7 +47,6 @@ func TestProcessor_AddMessage(t *testing.T) {
 }
 
 func TestProcessor_GetReply(t *testing.T) {
-	defer log.AfterTest(t)
 	p := NewServiceProcessor(nil)
 	log.ErrFatal(p.RegisterMessage(procMsg))
 
@@ -75,7 +73,6 @@ func TestProcessor_GetReply(t *testing.T) {
 }
 
 func TestProcessor_ProcessClientRequest(t *testing.T) {
-	defer log.AfterTest(t)
 	local := NewLocalTest()
 
 	// generate 5 hosts, they don't connect, they process messages, and they
