+++ sda/protocol_test.go
@@ -93,9 +93,6 @@ func TestProtocolRegistration(t *testing.T) {
 // and start a protocol. H1 should receive that message and request the entitity
 // list and the treelist and then instantiate the protocol.
 func TestProtocolAutomaticInstantiation(t *testing.T) {
-	defer log.AfterTest(t)
-
-	log.TestOutput(testing.Verbose(), 4)
 	// setup
 	chanH1 := make(chan bool)
 	chanH2 := make(chan bool)
