+++ services/identity/service_test.go
@@ -3,12 +3,24 @@ package identity
 import (
 	"testing"
 
+	"flag"
+	"os"
+
 	"github.com/dedis/cothority/log"
 	"github.com/dedis/cothority/network"
 	"github.com/dedis/cothority/sda"
 	"github.com/dedis/crypto/config"
 	"github.com/stretchr/testify/assert"
 )
+
+func TestMain(m *testing.M) {
+	//log.MainTest(m)
+	flag.Parse()
+	log.SetDebugVisible(1)
+	code := m.Run()
+	log.AfterTest(nil)
+	os.Exit(code)
+}
 
 func TestService_AddIdentity(t *testing.T) {
 	local := sda.NewLocalTest()
